requirements.txt
numpy
pandas
scikit-learn
tqdm
python-dotenv

opencv-python
Pillow
scikit-image
albumentations

torch
torchvision
torchaudio
timm

matplotlib
seaborn
grad-cam

langchain
langchain-community
langchain-text-splitters
faiss-cpu
sentence-transformers
transformers
accelerate

pypdf
pdfplumber
unstructured
unstructured[pdf]

joblib
pickle-mixin

fastapi
uvicorn
pydantic

requests
huggingface-hub
